Rename test dir
