+++ traffic_monitor/tests/_integration/README.md
@@ -7,4 +7,4 @@ From the `trafficcontrol/traffic_monitor` directory run `build_tests.sh`
 ## Running
 From the `trafficcontrol/traffic_monitor` directory run:
 
-`sudo docker-compose -p tmi --project-directory . -f tests/integration/docker-compose.yml run tmintegrationtest`
+`sudo docker-compose -p tmi --project-directory . -f tests/_integration/docker-compose.yml run tmintegrationtest`
