+++ traffic_monitor/tests/_integration/client_test.go
@@ -1,4 +1,4 @@
-package integration
+package _integration
 
 /*
    Licensed under the Apache License, Version 2.0 (the "License");
