+++ traffic_monitor/tests/_integration/docker-compose.yml
@@ -83,7 +83,7 @@ services:
 
   tmintegrationtest:
     build:
-      context: ./tests/integration
+      context: ./tests/_integration
       dockerfile: ./Dockerfile
     depends_on:
       - testto
