+++ traffic_monitor/tests/_integration/kbps_test.go
@@ -1,4 +1,4 @@
-package integration
+package _integration
 
 /*
